Use double.

+++ EmdFlat/Callback.cs
@@ -2,5 +2,5 @@
 
 namespace EmdFlat
 {
-    public delegate float Callback<feature_t>(feature_t item1, feature_t item2);
+    public delegate double Callback<feature_t>(feature_t item1, feature_t item2);
 }
