+++ EmdFlat/signature_t.cs
@@ -6,9 +6,9 @@ namespace EmdFlat
     {
         public int n;                 /* Number of features in the signature */
         public feature_t[] Features;  /* Pointer to the features vector */
-        public float[] Weights;       /* Pointer to the weights of the features */
+        public double[] Weights;      /* Pointer to the weights of the features */
 
-        public signature_t(int n, feature_t[] Features, float[] Weights)
+        public signature_t(int n, feature_t[] Features, double[] Weights)
         {
             this.n = n;
             this.Features = Features;
