+++ Workspace/Program.cs
@@ -6,7 +6,7 @@ public static class Program
 {
     public static unsafe void Main(string[] args)
     {
-        float[][] cost =
+        double[][] cost =
         [
             [3, 5, 2],
             [0, 2, 5],
@@ -17,8 +17,8 @@ public static class Program
 
         int[] f1 = [0, 1, 2, 3, 4];
         int[] f2 = [0, 1, 2];
-        float[] w1 = [0.4F, 0.2F, 0.2F, 0.1F, 0.1F];
-        float[] w2 = [0.6F, 0.2F, 0.1F];
+        double[] w1 = [0.4, 0.2, 0.2, 0.1, 0.1];
+        double[] w2 = [0.6, 0.2, 0.1];
         var s1 = new signature_t<int>(5, f1, w1);
         var s2 = new signature_t<int>(3, f2, w2);
 
@@ -26,7 +26,7 @@ public static class Program
         int flowSize = 0;
 
         var emd = new Emd();
-        float value;
+        double value;
         fixed (flow_t* p = flow)
         {
             value = emd.emd(s1, s2, (x, y) => cost[x][y], p, &flowSize);
